requirements.txt
streamlit==1.40.0
python-dotenv==1.0.1
pydantic==2.10.6
boto3==1.36.11
jinja2==3.1.5
googlesearch-python==1.3.0
pymupdf==1.25.2
duckduckgo-search==7.3.0
google-api-python-client==2.160.0
pandas==2.2.3
streamlit-folium==0.24.0
plotly==6.0.0